_backend_deprecated/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
pydantic
